Split pybind11 code in fem.cpp into two files.
Fixes #2351.
